backend/requirements.txt
fastapi
uvicorn
sqlalchemy
textblob
google-generativeai
python-dotenv
google-auth
requests